Turn redis-cart into a multi-cluster service
The redis-cart is the only stateful Service.
Because it's stateful, we only deploy it to a single cluster — the
cluster in the US.

multi-cluster-redis.yaml should be deployed to the the US cluster. It
contains a ServiceExport that basically says "export the redis-cart
Service to the other clusters".

The Terraform google_project_iam_member is explained at
https://cloud.google.com/kubernetes-engine/docs/how-to/multi-cluster-services#enabling