Use redis-cart.cartservice:6379 for US cluster
It takes some time for the ServiceExport to “export” the US cluster’s
redis-cart as a multi-cluster Service.
So let’s make the US cluster refer to the redis-cart as a local Service
(in the same cluster).
Why? So while ServiceExport is still working on creating
redis-cart.cartservice.svc.clusterset.local, the US cluster will be
functional.
(The redis-cart in only in the US cluster.)